migrate to mmm.hsgp module
